Move exceptions out
